# Load `/sys/kernel/tracing/available_filter_functions` lazily (and speedup tests)