<!--
Please provide a description of your change below this comment.

Then please complete the checklist.

Useful contribution guidelines and tips are in docs/developers.md.
-->

I was wondering why unit tests are taking so long. After a bit of profiling, I discovered that most of the time is spent in reading `/sys/kernel/tracing/available_filter_functions`. The file is parsed in the `BPFtrace` constructor which is not strictly necessary, b/c we need to query the list for certain probe types only (and almost never in unit tests).

This changes the approach to read the list lazily by creating a wrapper for accessing `BPFtrace::traceable_functions_` which only loads the list when accessed for the first time.

The speedup is huge, especially in unit tests. See below for numbers on my local setup.

Before:
    
    Unit tests:
    [==========] 410 tests from 22 test suites ran. (852498 ms total)
    Runtime tests:
    [==========] 554 tests from 26 test cases ran. (528572 ms total)

After:

    Unit tests:
    [==========] 410 tests from 22 test suites ran. (41140 ms total)
    Runtime tests:
    [==========] 554 tests from 26 test cases ran. (301417 ms total)

The PR also contains a patch to pretty-print elapsed time of runtime tests.

##### Checklist

- [ ] Language changes are updated in `man/adoc/bpftrace.adoc` and if needed in `docs/reference_guide.md`
- [ ] User-visible and non-trivial changes updated in `CHANGELOG.md`
- [ ] The new behaviour is covered by tests